refactor: split uteke-server monolith into modules (#514)
Split 1,840-line main.rs into 4 focused files:
- types.rs: request/response structs, helpers, constants
- context.rs: ReqCtx, auth, CORS middleware
- handlers.rs: route() dispatcher (all endpoint logic)
- main.rs: CLI args, config loading, server startup

Zero behavior change. All endpoints return identical responses.
Closes #514
